+++ src/main/java/de/kosit/validationtool/impl/DefaultCheck.java
@@ -81,8 +81,8 @@ public class DefaultCheck implements Check {
         this.repository = new ScenarioRepository(this.contentRepository);
         this.repository.initialize(configuration);
         this.checkSteps = new ArrayList<>();
-        this.checkSteps.add(new CreateDocumentIdentificationAction());
         this.checkSteps.add(new DocumentParseAction());
+        this.checkSteps.add(new CreateDocumentIdentificationAction());
         this.checkSteps.add(new ScenarioSelectionAction(this.repository));
         this.checkSteps.add(new SchemaValidationAction());
         this.checkSteps.add(new SchematronValidationAction(this.contentRepository, this.conversionService));